predictive-algorithm/pyproject.toml
[project]
name = "predictive-algorithm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.9.0",
    "numpy>=2.4.0",
    "numpy-quaternion>=2024.0.13",
    "pyqt6>=6.10.1",
    "pyvista[jupyter]>=0.46.4",
    "pyvistaqt>=0.11.3",
]
